Fix getActiveOasisDexOrders

+++ bin/compile.ts
@@ -45,7 +45,7 @@ const writeFiles = (compileOutput, contract) => {
   mkdirp.sync(targetDir);
 
   if (fs.existsSync(`${targetBasePath}.abi`)) {
-    throw new Error(
+    console.warn(
       // tslint:disable-next-line:max-line-length
       `Contract name duplication detected: ${targetBasePath}.abi. Please make sure that every contract is uniquely named across all dirctories.`,
     );
@@ -94,9 +94,11 @@ export const compileGlob = (
 
   debug('Writing compilation results');
 
-  // Delete and recreate out/
-  rimraf.sync(solidityCompileTarget);
-  mkdirp.sync(solidityCompileTarget);
+  if (query === path.join(soliditySourceDirectory, '**', '*.sol')) {
+    // Delete and recreate out/
+    rimraf.sync(solidityCompileTarget);
+    mkdirp.sync(solidityCompileTarget);
+  }
 
   fs.writeFileSync(
     path.join(solidityCompileTarget, 'compilerResult.json'),
